refactor: package typo
